+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - match Dream change from `bypassModelIntegrityCheck` to `bypassDreamIntegrityChecks`
 - match Dream change to allow automatic OpenAPI generation from `delegatedAttribute` serialization of associated models
+- fix resource controller spec generator missing date and datetime in spec ensuring model owned by another user is not updated
+- resource controller spec generator supports array attributes
 
 ## 1.10.5